requirements.txt

Flask
pyserial


keyboard
requests